requirements.txt
Django
djangorestframework
PyPDF2
langchain
openai
langchain-community
python-dotenv
InstructorEmbedding
chromadb
pypdf
faiss-cpu